Small modification

+++ TP_python/generate.py
@@ -3,12 +3,13 @@ import random
 import os
 import pickle
 import cv2
+from tqdm import tqdm
 
 from image_process import image_random_transform
 from dataset import load_dataset
 from utlis import load_dir_images
 
-def gene_train_images(ref='reference', train='appr', type_image_ref='bmp', type_image_train='png', N=100):
+def gene_train_images(ref='reference', train='appr/origin', type_image_ref='bmp', type_image_train='png', N=100):
     '''
     Generate N training images based on all the reference images by selecting randomly
     Generate a TXT file recording labels related to each image file name 
@@ -21,10 +22,16 @@ def gene_train_images(ref='reference', train='appr', type_image_ref='bmp', type_
         type_image_base (str): type des images generees dans la base
         N (int): nombre d'images generees
     '''
-    # delete all images
-    for f in Path('./'+train).glob("*.*"):
-        os.remove(f)
-        
+    CHECK_FOLDER = os.path.isdir(train)
+    # If folder doesn't exist, then create it.
+    if not CHECK_FOLDER:
+        print('[INFO] FOLDER NOT EXIST, CREATING A NEW ONE...')
+        os.makedirs(train)
+    else:
+        # delete all images
+        for f in Path('./'+train).glob("*.*"):
+            os.remove(f)
+    
     # load reference images
     img_ref_dict = load_dir_images(dir_name = './'+ref, img_type=type_image_ref)
     img_ref_names = list(img_ref_dict.keys())
@@ -32,22 +39,38 @@ def gene_train_images(ref='reference', train='appr', type_image_ref='bmp', type_
     # Generate N training images based on all the reference images by selecting randomly
     labels = {}
     images = {}
-    for i in range(N):
+    for i in tqdm(range(N)):
         ref_filename = random.choice(img_ref_names)
         image = cv2.imread(ref_filename, cv2.IMREAD_GRAYSCALE)
         image_transformed = image_random_transform(image)
-        filename = './{:s}/mesure{:03d}.{:s}'.format(train, i+1, type_image_train)
-        images[filename] = image_transformed
-        labels[filename] = int(ref_filename[-7])
+        key = 'mesure{:04d}.{:s}'.format(i+1, type_image_train)
+        images[key] = image_transformed
+        labels[key] = int(ref_filename[-7])
+        
         # save the rotated image to disk
+        filename = './{:s}/mesure{:04d}.{:s}'.format(train, i+1, type_image_train)
         cv2.imwrite(filename, image_transformed)
+        
     a_file = open("./{:s}/labels.pkl".format(train), "wb")
     pickle.dump(labels, a_file)
     a_file.close()
         
+    print('[INFO] Generating Dataset.......')
     dataset = load_dataset(images, labels)
     return dataset 
 
+
+def load_files_to_dataset(img_path='appr/origin',  type_image_train='png', label_path='appr/origin', label_filename='labels.pkl'):
+    print('[INFO] Loading Resource Files....... (this may take some time)')
+    labels_file = open('./{:s}/{:s}'.format(label_path, label_filename), "rb")
+    labels = pickle.load(labels_file)
+    images = load_dir_images(dir_name = './'+img_path, img_type=type_image_train, grayscale=True)
+    slice_keys = [w[-14:] for w in list(images.keys())]
+    new_images = dict(zip(slice_keys, list(images.values())))
+    
+    print('[INFO] Generating Dataset.......')
+    dataset = load_dataset(new_images, labels)
+    return dataset
 
 if __name__ == '__main__':
     dataset = gene_train_images()
